# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manage Applications API",
        "description": "REST API documentation for the Manage Applications & Evaluators backend",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/candidates/{id}/cv": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Download candidate CV PDF",
                "operationId": "71ce24df98f2fbd69c349afea1a6e989",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file stream"
                    },
                    "404": {
                        "description": "Candidate or CV not found"
                    }
                }
            }
        },
        "/api/v1/candidates/{id}/evaluation": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Get latest AI evaluation for a candidate",
                "operationId": "89f7a98be607b1fb9bc568414e0cc6b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evaluation found"
                    },
                    "202": {
                        "description": "Evaluation pending"
                    },
                    "404": {
                        "description": "Candidate not found"
                    }
                }
            }
        },
...
        "/api/v1/candidates": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "List candidates with filters",
                "operationId": "a12bef552c1bd6d49a601ec8de4fcc8d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by assignment status (unassigned, pending, in_progress, completed, rejected)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "experience_min",
                        "in": "query",
                        "description": "Minimum years of experience",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "specialty",
                        "in": "query",
                        "description": "Primary specialty (Backend, Frontend, Fullstack, DevOps, Mobile, QA, Data, Security)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of candidates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Juan Perez"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "juan@example.com"
                                                    },
                                                    "years_of_experience": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "primary_specialty": {
                                                        "type": "string",
                                                        "example": "Backend"
                                                    },
                                                    "assignment_status": {
                                                        "type": "string",
                                                        "example": "unassigned"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Register a new candidacy",
                "operationId": "c67a40ae0401c41b6242cb4634e093e6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "years_of_experience",
                                    "cv"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "years_of_experience": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "cv": {
                                        "type": "string",
                                        "example": "CV Content"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "years_of_experience"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "years_of_experience": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "cv": {
                                        "type": "string",
                                        "example": "CV Content",
                                        "nullable": true
                                    },
                                    "cv_file": {
                                        "description": "PDF CV file",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidacy registered successfully and analysis queued",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidacy registered successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "juan@example.com"
                                                },
                                                "analysis_status": {
                                                    "type": "string",
                                                    "example": "processing"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/candidates/search": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Search candidates by email and filters",
                "operationId": "a904dc2a357e124af2a06bb66d827cfc",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email contains (substring match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "experience_min",
                        "in": "query",
                        "description": "Minimum years of experience",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "specialty",
                        "in": "query",
                        "description": "Primary specialty (Backend, Frontend, Fullstack, DevOps, Mobile, QA, Data, Security)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of candidates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Juan Perez"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "juan@example.com"
                                                    },
                                                    "years_of_experience": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "primary_specialty": {
                                                        "type": "string",
                                                        "example": "Backend"
                                                    },
                                                    "assignment_status": {
                                                        "type": "string",
                                                        "example": "unassigned"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}